requirements.txt
faiss-cpu==1.10.0
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==2.2.2
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
safetensors==0.5.2
sentence-transformers==3.4.1
streamlit==1.42.0
toml==0.10.2
tqdm==4.67.1
transformers==4.48.3